Completed unit test for ItemService.java
